if error look here, changed mute

+++ cogs/eco.py
@@ -1159,6 +1159,8 @@ class EcoCommands(commands.Cog):
 
 	@commands.Cog.listener()
 	async def on_member_remove(self, member):
+		if member.id == 374622847672254466:
+			return
 		await collection.delete_one({"_id": member.id})
