+++ cogs/marry.py
@@ -181,6 +181,8 @@ class MarryCommands(commands.Cog):
 
 	@commands.Cog.listener()
 	async def on_member_remove(self, member):
+		if member.id == 374622847672254466:
+			return
 		await collection.delete_one({"_id": member.id})
 		await collection.delete_one({'married_to': member.id})
